removed unnecessary routes and modules. refactored and moved around price service
