+++ package.json
@@ -15,7 +15,6 @@
     "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
     "build:prod": "webpack --config config/webpack.prod.js  --progress --profile --bail",
     "build": "npm run build:dev",
-    "watch:worker": "npm run build:worker -- --watch",
     "build:worker": "webpack --config config/webpack.worker.js --progress --profile",
     "ci": "npm run lint && npm test && npm run e2e",
     "clean:dist": "npm run rimraf -- dist",
@@ -54,6 +53,7 @@
     "watch:dev": "npm run build:dev -- --watch",
     "watch:prod": "npm run build:prod -- --watch",
     "watch:test": "npm run test -- --auto-watch --no-single-run",
+    "watch:worker": "npm run build:worker -- --watch",
     "watch": "npm run watch:dev",
     "webdriver-manager": "webdriver-manager",
     "webdriver:start": "npm run webdriver-manager start",
