+++ src/app/app.routes.ts
@@ -1,5 +1,4 @@
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './home';
 import { AboutComponent } from './about';
 import { NoContentComponent } from './no-content';
 
@@ -9,7 +8,6 @@ import { StressComponent } from "./stress/stress.component";
 
 export const ROUTES: Routes = [
   { path: '',      component: StressComponent },
-  { path: 'home',  component: HomeComponent },
   { path: 'about', component: AboutComponent },
   { path: 'stress', component: StressComponent },
   {
