--- src/app/home/home.component.css
@@ -1,1 +1,0 @@
-/*styles for home content only*/
