--- src/app/home/home.component.html
@@ -1,1 +1,0 @@
-<tile-container [class.dark-theme]=true></tile-container>
