--- src/app/home/tile-container/index.ts
@@ -1,1 +1,0 @@
-export * from './tile-container.component';
