--- src/app/home/tile-container/tile-container.component.css
@@ -1,13 +1,0 @@
-.container{
-	display: flex;
-	flex-direction: row;
-	flex-wrap: wrap;
-	justify-content: flex-start;
-}
-
-[md-button].md-primary {
-	background-color: #E0E0E0;
-	margin: 5px;
-	border-radius: 15px;
-}
-
