--- src/app/home/tile-container/tile-container.component.html
@@ -1,23 +1,0 @@
-<div>
-	<button md-button (click)="start()" color="primary">Start</button>
-	<button md-button (click)="startWorker()" color="primary">Start Worker</button>
-	<button md-button (click)="stop()" color="primary">Stop</button>
-	<button md-button (click)="addPair()" color="primary">Add all pairs</button>
-	<button md-button (click)="clearAll()" color="primary">Clear</button>
-	<button *ngFor="let currencyPair of currencyPairs" md-button (click)="addPair(currencyPair)" color="primary">
-		{{ currencyPair }}
-	</button>
-
-	<label>Total Tiles : {{ state.selectedPairs.length }}</label>
-	<label>Update frequency: {{ frequency  }} ms</label>
-
-	<div class="container">
-		<tile *ngFor="let selectedPair of state.selectedPairs"
-		      [price1]="selectedPair.price1"
-		      [price2]="selectedPair.price2"
-		      [currency1]="selectedPair.currency1"
-		      [currency2]="selectedPair.currency2"
-		      [displayName]="selectedPair.key">
-		</tile>
-	</div>
-</div>
