--- src/app/home/tile/index.ts
@@ -1,1 +1,0 @@
-export * from './tile.component';
