--- src/app/home/tile/tile.component.css
@@ -1,23 +1,0 @@
-
-md-card .tile-sell{
-	background-color: #FFCDD2;
-	padding:15px 10px 4px 10px;
-	text-align: center;
-	border-radius: 15px;
-}
-
-md-card .tile-buy{
-	background-color: rgb(220,237,200);
-	padding:15px 10px 4px 10px;
-	text-align: center;
-	border-radius: 15px;
-}
-
-.tile-container{
-	padding: 5px;
-	text-align: center;
-}
-
-md-card {
-	margin: 4px;
-}
