--- src/app/home/tile/tile.component.html
@@ -1,23 +1,0 @@
-<md-card class.dark-theme="true">
-	<md-card-content>
-		<md-card-title-group>
-			<div class="tile-container">
-				<md-card-subtitle>{{ price1 }}</md-card-subtitle>
-				<div class="tile-sell">
-					<md-card-title>{{ price1 }}</md-card-title>
-					<md-card-title>Sell {{ currency1 }}</md-card-title>
-				</div>
-				<!--<md-slider min="1" max="5" step="0.5" value="1.5"></md-slider>-->
-			</div>
-			<div class="tile-container">
-				<md-card-subtitle>{{ price2 }}</md-card-subtitle>
-				<div class="tile-buy">
-					<md-card-title>{{ price2 }}</md-card-title>
-					<md-card-title>Buy {{ currency2 }}</md-card-title>
-				</div>
-				<!--<md-slider min="1" max="5" step="0.5" value="1.5"></md-slider>-->
-			</div>
-		</md-card-title-group>
-
-	</md-card-content>
-</md-card>
