--- src/app/home/x-large/index.ts
@@ -1,1 +1,0 @@
-export * from './x-large.directive';
