+++ src/app/stress/stress.component.css
@@ -1,1 +1,1 @@
-/*styles for home content only*/
+/*styles for stress content only*/
